+++ src/main/java/com/munecting/api/domain/user/constant/SocialType.java
@@ -1,4 +1,4 @@
-package com.munecting.api.domain.user.dto;
+package com.munecting.api.domain.user.constant;
 
 public enum SocialType {
     GOOGLE,KAKAO,NAVER,SPOTIFY,EMAIL
